requirements.txt
flask==2.3.2
tweepy==4.14.0
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.24.3